Rename componentExporter to componentLoader
